+++ application/controllers/Mtce.php
@@ -13,6 +13,101 @@ class Mtce extends Application {
             $this->page(1);
         }
 
+        // Initiate adding a new task (10.2)
+        public function add()
+        {
+            $task = $this->tasks->create();
+            $this->session->set_userdata('task', $task);
+            $this->showit();
+        }
+
+        // initiate editing of a task (10.3)
+        public function edit($id = null)
+        {
+            if ($id == null)
+                redirect('/mtce');
+            $task = $this->tasks->get($id);
+            $this->session->set_userdata('task', $task);
+            $this->showit();
+        }
+
+        // Render the current DTO (10.4)
+        private function showit()
+        {
+            $this->load->helper('form');
+            $task = $this->session->userdata('task');
+            $this->data['id'] = $task->id;
+
+            // if no errors, pass an empty message
+            if ( ! isset($this->data['error']))
+                $this->data['error'] = '';
+
+            $fields = array(
+                'ftask'      => form_label('Task description') . form_input('task', $task->task),
+                'fpriority'  => form_label('Priority') . form_dropdown('priority', $this->app->priority(), $task->priority),
+                'zsubmit'    => form_submit('submit', 'Update the TODO task'),
+            );
+            $this->data = array_merge($this->data, $fields);
+
+            $this->data['pagebody'] = 'itemedit';
+            $this->render();
+        }
+
+    // handle form submission (10.5)
+    public function submit()
+    {
+        // setup for validation
+        $this->load->library('form_validation');
+        $this->form_validation->set_rules($this->tasks->rules());
+
+        // retrieve & update data transfer buffer
+        $task = (array) $this->session->userdata('task');
+        $task = array_merge($task, $this->input->post());
+        $task = (object) $task;  // convert back to object
+        $this->session->set_userdata('task', (object) $task);
+
+        // validate away
+        if ($this->form_validation->run())
+        {
+            if (empty($task->id))
+            {
+                $task->id = $this->tasks->highest() + 1;
+                $this->tasks->add($task);
+                $this->alert('Task ' . $task->id . ' added', 'success');
+            } else
+            {
+                $this->tasks->update($task);
+                $this->alert('Task ' . $task->id . ' updated', 'success');
+            }
+        } else
+        {
+            $this->alert('<strong>Validation errors!<strong><br>' . validation_errors(), 'danger');
+        }
+        $this->showit();
+    }
+
+    // build a suitable error message (10.5)
+    private function alert($message) {
+        $this->load->helper('html');
+        $this->data['error'] = heading($message,3);
+    }
+
+    // Forget about this edit (10.6)
+    function cancel() {
+        $this->session->unset_userdata('task');
+        redirect('/mtce');
+    }
+
+    // Delete this item altogether (10.7)
+    function delete()
+    {
+        $dto = $this->session->userdata('task');
+        $task = $this->tasks->get($dto->id);
+        $this->tasks->delete($task->id);
+        $this->session->unset_userdata('task');
+        redirect('/mtce');
+    }
+
         
         private function show_page($tasks)
         {
@@ -54,6 +149,7 @@ class Mtce extends Application {
                     }
                     if ($count >= $this->items_per_page) break;
             }
+            $this->data['pagination'] = $this->pagenav($num);
             // INSERT next three lines
             $role = $this->session->userdata('userrole');
             if ($role == ROLE_OWNER)