+++ changelog.md
@@ -4,6 +4,25 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+##[1.6.0]
+###Changed
+ - fixed bug in models/app
+ - highest method in core/memory_model
+ 
+ 
+ ###Added
+ - task validation tools
+ - add function in mtce
+ - edit function in mtce
+ - showit function in mtce
+ - itemedit.php
+ - cancel button in itemedit.php
+ - deletion in itemedit.php
+ - form submission in mtce
+ - alert in mtce
+ - delete in mtce
+ 
+
 ## [1.5.0] - 2017-10-19
 ### Added
 - Role-specific maintenance list